move AdminPreview to own file
